refactor(Sample.Core): nest SampleCodeDownloader as Sample.Core.Downloader; split companion types; rename file
### Type relocation

- SampleCodeDownloader (file-scope class) -> Sample.Core.Downloader
- SampleDownloaderError (file-scope enum)  -> Sample.Core.Downloader.Error

### One type per file (task #21)

- Sample.Core.Downloader.swift           — main class + private auth helpers
- Sample.Core.Downloader.Error.swift     — extracted from file scope
- Sample.Core.Statistics.swift           — extracted from same file
- Sample.Core.Progress.swift             — extracted from same file

Was 1 file with 4 mixed types; now 4 single-type files.

### External caller sweep

- CLI/Command.Fetch.swift                    : 1 call site
- Tests/CoreTests/SampleCodeDownloaderTests.swift, MalformedURLSkipTests, CatalogTests, AuthPolicyTests

All references qualified through Sample.Core.Downloader.

### Imports added

- SampleCodeAuthPolicyTests.swift gained import SharedConstants (Sample namespace lives there).
- Sample.Core.{Statistics,Progress,Downloader.Error}.swift import SharedConstants.

### Verification

- xcrun swift build (clean): 0 warnings, 0 errors
- xcrun swift test (clean):  1300/1300 pass
