+++ Packages/Tests/CoreTests/SampleCodeAuthPolicyTests.swift
@@ -3,32 +3,33 @@ import AppKit
 @testable import Core
 import CoreProtocols
 import Foundation
+import SharedConstants
 import Testing
 
 /// Regression guard for #6. The authentication window only renders when the
 /// CLI process runs with `.regular` activation policy; `.prohibited` (the
 /// default for bare CLI tools) silently drops `NSWindow.makeKeyAndOrderFront`.
-@Suite("SampleCodeDownloader auth flow activation policy")
+@Suite("Sample.Core.Downloader auth flow activation policy")
 struct SampleCodeAuthPolicyTests {
     @Test("Auth flow requires .regular activation policy (fix for #6)")
     func authFlowPolicyIsRegular() {
-        #expect(SampleCodeDownloader.authFlowActivationPolicy == .regular)
+        #expect(Sample.Core.Downloader.authFlowActivationPolicy == .regular)
     }
 
     @Test("Auth flow policy is not .prohibited (the silent-no-op default)")
     func authFlowPolicyIsNotProhibited() {
-        #expect(SampleCodeDownloader.authFlowActivationPolicy != .prohibited)
+        #expect(Sample.Core.Downloader.authFlowActivationPolicy != .prohibited)
     }
 
     @Test("Auth flow policy is not .accessory (menu-bar-only, no window)")
     func authFlowPolicyIsNotAccessory() {
-        #expect(SampleCodeDownloader.authFlowActivationPolicy != .accessory)
+        #expect(Sample.Core.Downloader.authFlowActivationPolicy != .accessory)
     }
 }
 
 // MARK: - Cookie detection (#6 follow-up auto-advance)
 
-@Suite("SampleCodeDownloader.containsAppleSessionCookie")
+@Suite("Sample.Core.Downloader.containsAppleSessionCookie")
 struct AppleSessionCookieDetectionTests {
     private static func cookie(name: String, domain: String) -> HTTPCookie {
         HTTPCookie(properties: [
@@ -41,31 +42,31 @@ struct AppleSessionCookieDetectionTests {
 
     @Test("Empty cookie list returns false")
     func emptyList() {
-        #expect(SampleCodeDownloader.containsAppleSessionCookie([]) == false)
+        #expect(Sample.Core.Downloader.containsAppleSessionCookie([]) == false)
     }
 
     @Test("myacinfo on apple.com domain counts as signed-in")
     func myacinfoOnAppleCom() {
         let cookies = [Self.cookie(name: "myacinfo", domain: ".apple.com")]
-        #expect(SampleCodeDownloader.containsAppleSessionCookie(cookies) == true)
+        #expect(Sample.Core.Downloader.containsAppleSessionCookie(cookies) == true)
     }
 
     @Test("myacinfo on developer.apple.com subdomain also counts")
     func myacinfoOnDeveloperSubdomain() {
         let cookies = [Self.cookie(name: "myacinfo", domain: "developer.apple.com")]
-        #expect(SampleCodeDownloader.containsAppleSessionCookie(cookies) == true)
+        #expect(Sample.Core.Downloader.containsAppleSessionCookie(cookies) == true)
     }
 
     @Test("myacinfo on a non-apple domain does NOT count")
     func myacinfoOnUnrelatedDomain() {
         let cookies = [Self.cookie(name: "myacinfo", domain: "example.com")]
-        #expect(SampleCodeDownloader.containsAppleSessionCookie(cookies) == false)
+        #expect(Sample.Core.Downloader.containsAppleSessionCookie(cookies) == false)
     }
 
     @Test("Unknown cookie name on apple.com does NOT count")
     func unrelatedCookieOnAppleDomain() {
         let cookies = [Self.cookie(name: "random_marketing", domain: ".apple.com")]
-        #expect(SampleCodeDownloader.containsAppleSessionCookie(cookies) == false)
+        #expect(Sample.Core.Downloader.containsAppleSessionCookie(cookies) == false)
     }
 
     @Test("Mix of unrelated cookies plus the target cookie counts")
@@ -75,58 +76,58 @@ struct AppleSessionCookieDetectionTests {
             Self.cookie(name: "tracking", domain: "example.com"),
             Self.cookie(name: "myacinfo", domain: "idmsa.apple.com"),
         ]
-        #expect(SampleCodeDownloader.containsAppleSessionCookie(cookies) == true)
+        #expect(Sample.Core.Downloader.containsAppleSessionCookie(cookies) == true)
     }
 
     @Test("Case-insensitive apple.com matching (APPLE.COM)")
     func caseInsensitiveDomain() {
         let cookies = [Self.cookie(name: "myacinfo", domain: "APPLE.COM")]
-        #expect(SampleCodeDownloader.containsAppleSessionCookie(cookies) == true)
+        #expect(Sample.Core.Downloader.containsAppleSessionCookie(cookies) == true)
     }
 
     @Test("Target cookie name set includes myacinfo")
     func targetCookieNameSet() {
-        #expect(SampleCodeDownloader.appleSessionCookieNames.contains("myacinfo"))
+        #expect(Sample.Core.Downloader.appleSessionCookieNames.contains("myacinfo"))
     }
 }
 
 // MARK: - TTY detection (#6 follow-up)
 
-@Suite("SampleCodeDownloader.isInteractiveStdin")
+@Suite("Sample.Core.Downloader.isInteractiveStdin")
 struct IsInteractiveStdinTests {
     @Test("Override seam returns whatever was forced")
     func overrideIsRespected() {
-        let previous = SampleCodeDownloader._isInteractiveStdinOverride
-        defer { SampleCodeDownloader._isInteractiveStdinOverride = previous }
+        let previous = Sample.Core.Downloader._isInteractiveStdinOverride
+        defer { Sample.Core.Downloader._isInteractiveStdinOverride = previous }
 
-        SampleCodeDownloader._isInteractiveStdinOverride = true
-        #expect(SampleCodeDownloader.isInteractiveStdin() == true)
+        Sample.Core.Downloader._isInteractiveStdinOverride = true
+        #expect(Sample.Core.Downloader.isInteractiveStdin() == true)
 
-        SampleCodeDownloader._isInteractiveStdinOverride = false
-        #expect(SampleCodeDownloader.isInteractiveStdin() == false)
+        Sample.Core.Downloader._isInteractiveStdinOverride = false
+        #expect(Sample.Core.Downloader.isInteractiveStdin() == false)
     }
 
     @Test("Clearing the override falls back to the real isatty check")
     func noOverrideFallsBack() {
-        let previous = SampleCodeDownloader._isInteractiveStdinOverride
-        defer { SampleCodeDownloader._isInteractiveStdinOverride = previous }
+        let previous = Sample.Core.Downloader._isInteractiveStdinOverride
+        defer { Sample.Core.Downloader._isInteractiveStdinOverride = previous }
 
-        SampleCodeDownloader._isInteractiveStdinOverride = nil
+        Sample.Core.Downloader._isInteractiveStdinOverride = nil
         // We don't assert true/false here — swift test's stdin may or may not
         // be a TTY depending on how the runner invoked us. Just confirm the
         // call does not crash and returns a Bool.
-        _ = SampleCodeDownloader.isInteractiveStdin()
+        _ = Sample.Core.Downloader.isInteractiveStdin()
     }
 }
 
 // MARK: - AuthOutcome enum
 
-@Suite("SampleCodeDownloader.AuthOutcome")
+@Suite("Sample.Core.Downloader.AuthOutcome")
 struct AuthOutcomeTests {
     @Test("AuthOutcome exposes exactly the three expected cases")
     func outcomeCases() {
         // Exhaustive switch pins the case list at compile time.
-        let all: [SampleCodeDownloader.AuthOutcome] = [
+        let all: [Sample.Core.Downloader.AuthOutcome] = [
             .autoDetected, .userConfirmed, .userClosedWindow,
         ]
         #expect(all.count == 3)
